+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## RedVox Pandas (RedPandas) Version History
 
+## 1.3.0 ()
+- Backend changes to adapt to RedVox SDK 3.1.0
+- Added RedPandas version in the RedPandas DataFrame
+
 ## 1.2.13-15 (2021-10-28)
 - Minor changes to mesh plot and decimating functions
 - Changes to mesh to avoid memory leaks